finally going in to the colour cells for loop, no longer empty
i assigned a new variable of taskcells to hold the generated spawn cells. the spawncell method now returns it at the end. This then is passed into the colour cells which now has a full list, hence the for loop works- sort of, now we have a new error saying method is not subscriptable, i think its maybe a brackets problem